Handle TLS configuration correctly
Currently, TLS is only configured if all of CA, client certificate and
client key are provided. This also includes `tlsSkipVerify`, which makes
it mostly useless as it is only respected if you also provide a CA and
client cert.

In addition, it's currently impossible to specify just the CA without
setting a client certificate.

We can fix this by setting only the configuration provided, which is
also in line with how `nomad` behaves.