Disambiguate the names of the Rust client crates.
This is a breaking change, as it renames the two `nomad_client` crates and thus every symbol exported by either of them.

Cargo's `git` dependency resolution is able to locate crates in subdirectories of a Git repository, so long as the crate names are unique within the repo[1]. Unfortunately, the existing crates are both named `nomad_client`, which is ambiguous and causes Cargo to arbitrarily pick one when resolving the crate.

This change distinguishes the crates so that they have distinct names, allowing clients to pick the crate and depend on it directly from Git. The option to vendor the crate remains open, but it's no longer strictly required, which simplifies dependency upgrades a bit (`cargo upgrade` can update the commit hash, for example).

Publishing these crates to crates.io for general use is beyond the scope of this change.

[1]: https://github.com/rust-lang/cargo/issues/1462